powerpc: turris: Fix addrs for SDCARD SRAM boot
Fix CONFIG_SYS_TEXT_BASE and CONFIG_RESET_VECTOR_ADDRESS for
SDCARD target. The board uses multi-stage boot with ROM code
for loading the U-Boot binary to certain address either in
DRAM or SRAM (when it is large enough). Turris1 board uses
SRAM because it is more dependable and therefore the addresses
has to be adjusted accordingly.